requirements.txt
streamlit
torch
torchvision
timm
huggingface_hub
grad-cam
Pillow
numpy
matplotlib
